pyproject.toml
[project]
name = "viberl"
version = "0.1.0"
description = "A toolbox for prototyping new RL approaches"
readme = "README.md"
authors = [{ name = "Naren Dasan", email = "naren@narendasan.com" }]
requires-python = ">=3.11"
dependencies = [
    "jax[cuda12]>=0.6.0; sys_platform == 'linux'",
    "jax>=0.6.0; sys_platform == 'darwin'",
    #"jax-metal; sys_platform == 'darwin'",
    "orbax-checkpoint>=0.11.10",
    "rejax>=0.1.1",
    "tomli-w>=1.2.0",
    "ribs>=0.7.1",
    "distrax>=0.1.5",
    "flax>=0.10.6",
    "gymnax @ git+https://github.com/smorad/stable-gymnax@main",
    "xxhash",
]
license = "NCSA"

[project.optional-dependencies]
wandb = ["wandb>=0.19.11"]
mlflow = ["mlflow>=2.20.2"]

[dependency-groups]
dev = ["isort>=6.0.0", "ruff>=0.9.5", "marimo>=0.11.5", "pre-commit>=2.20.0"]

[tool.uv]
package = true
#default-groups = ["dev"]

[tool.uv.sources]
rejax = { path = "../rejax", editable = true }

[tool.ruff]
# NOTE: Synchoronize the ignores with .flake8
lint.ignore = [
    # these ignores are from flake8-bugbear; please fix!
    "B007",
    "B008",
    "B017",
    "B018",   # Useless expression
    "B019",
    "B020",
    "B023",
    "B024",
    "B026",
    "B028",   # No explicit `stacklevel` keyword argument found
    "B904",
    "B905",
    "E402",
    "C408",   # C408 ignored because we like the dict keyword argument syntax
    "E501",   # E501 is not flexible enough, we're using B950 instead
    "E721",
    "E731",   # Assign lambda expression
    "E741",
    "EXE001",
    "F405",
    "F821",
    "F841",
    # these ignores are from flake8-logging-format; please fix!
    "G101",
    "G201",
    "G202",
    "G003",
    "G004",
    # these ignores are from RUFF perf; please fix!
    "PERF203",
    "PERF4",
    "SIM102",
    "SIM103",
    "SIM112",  # flake8-simplify code styles
    "SIM105",  # these ignores are from flake8-simplify. please fix or ignore with commented reason
    "SIM108",
    "SIM110",
    "SIM114",  # Combine `if` branches using logical `or` operator
    "SIM115",
    "SIM116",  # Disable Use a dictionary instead of consecutive `if` statements
    "SIM117",
    "SIM118",
]
#line-length = 120
lint.select = [
    "B",
    "C4",
    "G",
    "E",
    "F",
    "SIM1",
    "W",
    # Not included in flake8
    "PERF",
    "PLE",
    "TRY302",
]

# Allow unused variables when underscore-prefixed.
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Allow autofix for all enabled rules (when `--fix`) is provided.
lint.fixable = [
    "A",
    "B",
    "C",
    "D",
    "E",
    "F",
    "G",
    "I",
    "N",
    "Q",
    "S",
    "T",
    "W",
    "ANN",
    "ARG",
    "BLE",
    "COM",
    "DJ",
    "DTZ",
    "EM",
    "ERA",
    "EXE",
    "FBT",
    "ICN",
    "INP",
    "ISC",
    "NPY",
    "PD",
    "PGH",
    "PIE",
    "PL",
    "PT",
    "PTH",
    "PYI",
    "RET",
    "RSE",
    "RUF",
    "SIM",
    "SLF",
    "TCH",
    "TID",
    "TRY",
    "UP",
    "YTT",
]
lint.unfixable = []
target-version = "py311"

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "env",
    "py/torch_tensorrt/fx",
    ".github",
    "examples",
    "tests",
    "tools",
    "docs",
    "docsrc",
    "tests",
    "setup.py",
    "noxfile.py",
    "__init__.py",
]

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.isort]
profile = "black"
py_version = 311
